+++ lib/paper_trail/serializers/json.rb
@@ -34,7 +34,7 @@ module PaperTrail
 
       def where_object_changes_condition(*)
         raise <<-STR.squish.freeze
-          where_object_changes no longer supports reading json from a text
+          where_object_changes no longer supports reading JSON from a text
           column. The old implementation was inaccurate, returning more records
           than you wanted. This feature was deprecated in 7.1.0 and removed in
           8.0.0. The json and jsonb datatypes are still supported. See the