+++ spec/models/document_spec.rb
@@ -9,17 +9,6 @@ RSpec.describe Document, type: :model, versioning: true do
     end
   end
 
-  describe "have_a_version_with_changes matcher" do
-    it "works with custom versions association" do
-      document = Document.create!(name: "Foo")
-      document.update_attributes!(name: "Bar")
-      allow(ActiveSupport::Deprecation).to receive(:warn)
-      expect(document).to have_a_version_with_changes(name: "Bar")
-      expect(ActiveSupport::Deprecation).to have_received(:warn).
-        once.with(/^where_object_changes/)
-    end
-  end
-
   describe "#paper_trail.next_version" do
     it "returns the expected document" do
       doc = Document.create