+++ spec/models/widget_spec.rb
@@ -21,24 +21,6 @@ RSpec.describe Widget, type: :model do
     end
   end
 
-  describe "`have_a_version_with_changes` matcher", versioning: true do
-    before do
-      widget.update_attributes!(name: "Leonard", an_integer: 2)
-      widget.update_attributes!(name: "Tom")
-      widget.update_attributes!(name: "Bob")
-    end
-
-    it "is possible to do assertions on version changes" do
-      allow(ActiveSupport::Deprecation).to receive(:warn)
-      expect(widget).to have_a_version_with_changes name: "Leonard", an_integer: 2
-      expect(widget).to have_a_version_with_changes an_integer: 2
-      expect(widget).to have_a_version_with_changes name: "Tom"
-      expect(widget).to have_a_version_with_changes name: "Bob"
-      expect(ActiveSupport::Deprecation).to have_received(:warn).
-        exactly(5).times.with(/^where_object_changes/)
-    end
-  end
-
   describe "versioning option" do
     context "enabled", versioning: true do
       it "enables versioning" do